Making it so dictionary keys are the English values for strings displayed in the UI.
By making dictionary keys be sensible strings, we make it so the i18n function returns something legible when there are no dictionaries included in a build.  This means that strings from the Graticule, Permalink, LayerSwitcher, and Scale controls will be displayed in English without having to include the English dictionary (as is the case in our full build).

Note that if you use the OpenLayers.i18n method in your applications, five of the dictionary keys have changed.  Use the following replacements:

 * "permalink" -> "Permalink"
 * "overlays" -> "Overlays"
 * "baseLayer" -> "Base Layer"
 * "scale" -> "Scale = 1 : ${scaleDenom}"
 * "graticule" -> "Graticule"

r=ahocevar (closes #3364)



git-svn-id: http://svn.openlayers.org/trunk@12127 dc9f47b5-9b13-0410-9fdd-eb0c1a62fdaf